fix(design-system): move ts generated files into src folder
