+++ .storybook/webpack.config.js
@@ -4,7 +4,7 @@ const MomentLocalesPlugin = require('moment-locales-webpack-plugin');
 const sourceFolders = [
   path.resolve(__dirname),
   path.resolve(__dirname, '../examples'),
-  path.resolve(__dirname, '../design-system/materials'),
+  path.resolve(__dirname, '../design-system'),
   path.resolve(__dirname, '../philosophy'),
   path.resolve(__dirname, '../packages'),
   path.resolve(__dirname, '../presets'),
