+++ design-system/src/custom-properties.ts
