+++ design-system/src/index.ts
@@ -1,4 +1,4 @@
-export { default as customProperties } from '../materials/custom-properties';
-export { default as designTokens } from '../materials/design-tokens';
+export { default as customProperties } from './custom-properties';
+export { default as designTokens } from './design-tokens';
 
 export { default as version } from './version';
